+++ src/ksi/net_tcp_async.c
@@ -430,7 +430,7 @@ static int dispatch(TcpAsyncCtx *tcpCtx) {
 
 				/* Check if the request count can be restarted. */
 				if (difftime(time(&curTime), tcpCtx->roundStartAt) >= tcpCtx->options[KSI_ASYNC_PRIVOPT_ROUND_DURATION]) {
-					KSI_LOG_info(tcpCtx->ctx, "Async TCP round request count: %u", tcpCtx->roundCount);
+					KSI_LOG_info(tcpCtx->ctx, "Async TCP round request count: %llu", (unsigned long long)tcpCtx->roundCount);
 					tcpCtx->roundCount = 0;
 					tcpCtx->roundStartAt = curTime;
 				}
@@ -474,7 +474,7 @@ static int dispatch(TcpAsyncCtx *tcpCtx) {
 								if (KSI_SCK_errno == KSI_SCK_EWOULDBLOCK || KSI_SCK_errno == KSI_SCK_EAGAIN) {
 									KSI_LOG_info(tcpCtx->ctx,
 											"Async TCP send would block. Bytes sent so far %d/%d. Error: %d (%s).",
-											req->sentCount, req->len, KSI_SCK_errno, KSI_SCK_strerror(KSI_SCK_errno));
+											(unsigned)req->sentCount, (unsigned)req->len, KSI_SCK_errno, KSI_SCK_strerror(KSI_SCK_errno));
 									goto cleanup;
 								} else {
 									closeSocket(tcpCtx, __LINE__);
